Move exportToCsv back to project.controller from project.entity
