+++ src/server/src/projects/entities/project.entity.ts
@@ -6,8 +6,7 @@ import {
   DistrictProperties,
   DistrictsDefinition,
   IProject,
-  MetricField,
-  GeoUnitHierarchy
+  MetricField
 } from "../../../../shared/entities";
 import { RegionConfig } from "../../region-configs/entities/region-config.entity";
 import { Chamber } from "../../chambers/entities/chamber.entity";
@@ -17,14 +16,9 @@ import {
   DEFAULT_POPULATION_DEVIATION,
   DEFAULT_PINNED_METRIC_FIELDS
 } from "../../../../shared/constants";
-import { GeoUnitTopology } from "../../districts/entities/geo-unit-topology.entity";
-import { InternalServerErrorException, Logger } from "@nestjs/common";
-import { GeoUnitProperties } from "../../districts/entities/geo-unit-properties.entity";
 
 // TODO #179: Move to shared/entities
 export type DistrictsGeoJSON = FeatureCollection<MultiPolygon, DistrictProperties>;
-
-const LOGGER = new Logger("Project");
 
 @Entity()
 export class Project implements IProject {
@@ -120,45 +114,5 @@ export class Project implements IProject {
   // Strips out data that we don't want to have available in the read-only view in the UI
   getReadOnlyView(): Project {
     return { ...this, lockedDistricts: new Array(this.numberOfDistricts).fill(false) };
-  }
-
-  exportToCsv(geoCollection: GeoUnitTopology | GeoUnitProperties): [string, number][] {
-    const baseGeoLevel = geoCollection.definition.groups.slice().reverse()[0];
-    const baseGeoUnitProperties = geoCollection.topologyProperties[baseGeoLevel];
-
-    // First column is the base geounit id, second column is the district id
-    const mutableCsvRows: [string, number][] = [];
-
-    // The geounit hierarchy and district definition have the same structure (except the
-    // hierarchy always goes out to the base geounit level). Walk them both at the same time
-    // and collect the information needed for the CSV (base geounit id and district id).
-    const accumulateCsvRows = (
-      defnSubset: number | GeoUnitHierarchy,
-      hierarchySubset: GeoUnitHierarchy
-    ) => {
-      hierarchySubset.forEach((hierarchyNumOrArray, idx) => {
-        const districtOrArray = typeof defnSubset === "number" ? defnSubset : defnSubset[idx];
-        if (typeof districtOrArray === "number" && typeof hierarchyNumOrArray === "number") {
-          // The numbers found in the hierarchy are the base geounit indices of the topology.
-          // Access this item in the topology to find it's base geounit id.
-          const props: any = baseGeoUnitProperties[hierarchyNumOrArray];
-          mutableCsvRows.push([props[baseGeoLevel], districtOrArray]);
-        } else if (typeof hierarchyNumOrArray !== "number") {
-          // Keep recursing into the hierarchy until we reach the end
-          accumulateCsvRows(districtOrArray, hierarchyNumOrArray);
-        } else {
-          // This shouldn't ever happen, and would suggest a district definition/hierarchy mismatch
-          LOGGER.error(
-            "Hierarchy and districts definition mismatch",
-            districtOrArray.toString(),
-            hierarchyNumOrArray.toString()
-          );
-          throw new InternalServerErrorException();
-        }
-      });
-    };
-    accumulateCsvRows(this.districtsDefinition, geoCollection.hierarchyDefinition);
-
-    return mutableCsvRows;
   }
 }
